UKF: Implement UKF Algorithm
This commit introduces the ukf algorithm. In the ukf class the predict and
the update functions are called. All the sensor specific parts are like the
measurment_functaion, the R_ matrix and the noise parameters are in the
specific sensor implementations. Now the measuremen_package class contains
a sensor field in which the specific Sensor is stored.